feat(editor): ✨ Adds basic entity editor
Basic entity editor allows for editing position and rotation of entities. Proves out the gameplay facade idea.

+++ triton/include/gfx/Handles.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 namespace tr::gfx {
-
-   using TextureHandle = size_t;
+   using TextureHandle = uint32_t;
    using MeshHandle = size_t;
 }
